requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
certifi==2025.8.3
cloudinary==1.44.1
distro==1.9.0
Django==4.2.23
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.1
openai==1.99.1
pillow==10.4.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
